Moved DurationFilter
